The Country and area classification according to the United Nations Standaistical Division, Standard country or area codes for statistical use (M49). Adapted for use in FP2030 by the Track20 project. A subset of data from the United Nations country classifications

The Country and area classification according to the United Nations Standaistical Division, Standard country or area codes for statistical use (M49). Adapted for use in FP2030 by the Track20 project. A subset of data from the United Nations country classifications

Usage

Country_and_area_classification

Format

A data frame with 231 rows and 8 columns:

Country or area

Country name

ISO Code

1, 2 & 3 number ISO country codes

Major area

Continent

Region

Sub-continent

Developed region

Binary indicator for development status

Least developed country

Binary indicator for least developed status

Sub-Saharan Africa

Binary indicator for whether country is in Sub-Saharan Africa

FP2020

Binary indicator for FP2020 participation status

Source

https://unstats.un.org/unsd/methodology/m49/