DHS survey observations for the proportion of modern contraceptives supplied by the public and private sectors at the subnational administration level.

Elizabeth Heger Boyle, Miriam King and Matthew Sobek. IPUMS-Demographic and Health Surveys: Version 9 (dataset). IPUMS and ICF, 2022. https://doi.org/10.18128/D080.V9

Usage

subnat_FPsource_data

Format

subnat_FPsource_data

A data frame with 6940 rows and 8 columns:

Country

Country names

Region

Subnational region names

Method

Contraceptive method name

average_year

Average year of the survey

sector_categories

Name of sector

proportion

Proportion supplied by the sector

SE.proportion

Standard error associated with the proportion

n

Sample size associated with the observation

Source

On request from IPUMS - https://www.idhsdata.org/idhs/index.shtml