Get JAGS model inputs

Get JAGS model inputs

get_modelinputs(startyear = 1990, endyear = 2030.5, nsegments = 12, raw_data)

Arguments

startyear

The year you wish to begin your predictions from. Default is 1990.

endyear

The year you wish to finish your predictions. Default is 2030.5.

nsegments

The number of knots you wish to use in your basis functions. Default is 12.

raw_data

The list of arguments and family planning source data from the 'get_data' function.

Value

A list of modelling inputs for the JAGS model.

Tstar is the year index for the most recent survey in each province.

Kstar is the knot index that aligns with Tstar.

B.ik are the basis functions.

n_years are total number of years

n_obs are the total number of observations

K are the number of knots.

H is K-1. Used in the calculation of first order differences of spline coefficients.

P_count is the number of subnational provinces/regions.

M_count is the number of modern contraceptive methods.

matchsubnat is the subnational province indexing to match the observed data to the predictions.

matchcountry is the country indexing to match the observed data to the predictions.

matchmethod is the method indexing to match the observed data to the predictions.

matchyears is the year indexing to match the observed data to the predictions.

Examples

# \donttest{
raw_data <- get_data(national=FALSE, local=TRUE, mycountry="Nepal")
#> Using preloaded dataset!
#> Joining with `by = join_by(Country, average_year)`
#> Getting data for Nepal
#> Joining with `by = join_by(Country)`
jagsdata <- get_modelinputs(startyear=1990, endyear=2030.5, nsegments=12, raw_data)
#> Adding missing grouping variables: `Country`, `index_country`
# }